refactor - cleaner and got rid of unnecessary code
